+++ Sources/Conduit/Networking/Serialization/XML/XMLNode.swift
@@ -127,6 +127,13 @@ public struct XMLNode {
     }
 }
 
+/// Traversal options for extracting nodes from an XMLNode tree
+public enum XMLNodeTraversal {
+    case firstLevel
+    case depthFirst
+    case breadthFirst
+}
+
 // MARK: CustomStringConvertible
 
 extension XMLNode: CustomStringConvertible {
@@ -193,9 +200,9 @@ public extension XMLNode {
     /// - Parameter traversal: Node Traversal technique. Defaults to Breadth first
     /// - Returns: Node value (text node) converted to given type
     /// - Throws: XMLError if no descendant found, node has no value (does not contain a text node)
-    ///           or if casting to type fails
-    public func getValue<T: LosslessStringConvertible>(_ name: String, traversal: XMLNodeTraversal = .breadthFirst) throws -> T {
-        return try node(named: name, traversal: traversal).getValue() as T
+    ///           and no default has been provided, or if casting to type fails
+    public func getValue<T: XMLTextNodeInitializable>(_ name: String, traversal: XMLNodeTraversal = .breadthFirst) throws -> T {
+        return try node(named: name, traversal: traversal).getValue()
     }
 
     /// Retrieve the first descendant node with the given name, converted to the given type
@@ -205,39 +212,67 @@ public extension XMLNode {
     /// - Returns: Node value (text node) converted to given type
     /// - Throws: XMLError if no descendant found, node has no value (does not contain a text node)
     ///           or if casting to type fails
-    public func getValue<T: LosslessStringConvertible>(_ name: String, traversal: XMLNodeTraversal = .breadthFirst) -> T? {
-        guard let node = try? node(named: name, traversal: traversal) else {
-            return nil
-        }
-        return try? node.getValue()
+    public func getValue<T: XMLTextNodeInitializable>(_ name: String, traversal: XMLNodeTraversal = .breadthFirst) -> T? {
+        return try? node(named: name, traversal: traversal).getValue()
     }
 
     /// Get node value (text node) converted to given type
     ///
     /// - Returns: Node value (text node) converted to given type
     /// - Throws: XMLError if node has no value (does not contain a text node) or if casting to type fails
-    public func getValue<T: LosslessStringConvertible>() throws -> T {
-        guard let result: T = self.getValue() else {
+    public func getValue<T: XMLTextNodeInitializable>() throws -> T {
+        guard let value: T = getValue() else {
             throw XMLError.invalidDataType
         }
-        return result
+        return value
     }
 
     /// Get node valye (text node) converted to a given type
     ///
     /// - Returns: Node value (text node) converted to a given type, otherwise nil
-    public func getValue<T: LosslessStringConvertible>() -> T? {
-        guard let text = self.text else {
+    public func getValue<T: XMLTextNodeInitializable>() -> T? {
+        guard let textNode = text else {
             return nil
         }
-        return T(text)
+        return T(xmlTextNode: textNode)
     }
 
 }
 
-/// Traversal options for extracting nodes from an XMLNode tree
-public enum XMLNodeTraversal {
-    case firstLevel
-    case depthFirst
-    case breadthFirst
+// MARK: - XMLTextNodeInitializable
+
+/// Types conforming to `XMLTextNodeInitializable` can be constructed directly
+/// from `XMLNode` text value.
+public protocol XMLTextNodeInitializable {
+    init?(xmlTextNode: String)
+}
+
+extension Int: XMLTextNodeInitializable {
+    public init?(xmlTextNode: String) {
+        self.init(xmlTextNode, radix: 10)
+    }
+}
+
+extension Double: XMLTextNodeInitializable {
+    public init?(xmlTextNode: String) {
+        self.init(xmlTextNode)
+    }
+}
+
+extension Decimal: XMLTextNodeInitializable {
+    public init?(xmlTextNode: String) {
+        self.init(string: xmlTextNode)
+    }
+}
+
+extension String: XMLTextNodeInitializable {
+    public init?(xmlTextNode: String) {
+        self = xmlTextNode
+    }
+}
+
+extension Bool: XMLTextNodeInitializable {
+    public init?(xmlTextNode: String) {
+        self.init(xmlTextNode)
+    }
 }